+++ src/core/topology.c
@@ -797,7 +797,7 @@ static void _starpu_initialize_workers_bindid(struct _starpu_machine_config *con
 							!hwloc_bitmap_isset(config->topology.log_coreset, val))
 							_STARPU_DISP("Warning: logical core id %ld is not in the CPU binding provided by the OS\n", val);
 #endif
-						if (val * scale >= topology->nhwpus)
+						if (val * scale >= (long int)topology->nhwpus)
 							_STARPU_DISP("Warning: logical core id %ld is beyond the number of cores (%d), will wrap around it\n", val, topology->nhwpus / scale);
 					}
 					else
@@ -807,7 +807,7 @@ static void _starpu_initialize_workers_bindid(struct _starpu_machine_config *con
 							!hwloc_bitmap_isset(config->topology.log_cpuset, val))
 							_STARPU_DISP("Warning: logical CPU id %ld is not in the CPU binding provided by the OS\n", val);
 #endif
-						if (val >= topology->nhwpus)
+						if (val >= (long int)topology->nhwpus)
 							_STARPU_DISP("Warning: logical CPU id %ld is beyond the number of CPUs (%d), will wrap around it\n", val, topology->nhwpus);
 					}
 					topology->workers_bindid[i] = (unsigned)((val * scale) % topology->nhwpus);
